+++ NEWS.md
@@ -11,6 +11,15 @@ I am successful in that undertaking.
 Consult [the Github repository](github.com/verumsolum/orf_weather) if you
 require details about earlier changes.
 
+### Changes in version 0.0.0.9037 (2017-01-09)
+* **computeCumulativePrecipitation**: Add `computeCumulativePrecipitation`
+function to calculate `MTDPrecip` and `YTDPrecip` variables
+* **computeCumulativePrecipRecords**: Find records (and years) for
+`MTDPrecip` and `YTDPrecip`
+* **plotMTDPrecipitation**: Plot month-to-date precipitation, with historical
+maximum and minimum month-to-date precipitation totals.
+* **%>%**: Import function from `dplyr`
+
 ### Changes in version 0.0.0.9036 (2017-01-06)
 * **getUpdatedCsv**: Use `utils::download.file` instead of `httr:` functions